Updates the jobs to use the new file structure, adds the remap_grid_name and native_grid_name keys to the config

+++ lib/runmanager.py
@@ -4,7 +4,7 @@ import time
 import re
 from datetime import datetime
 from threading import Thread
-from shutil import copytree
+from shutil import copytree, move
 from subprocess import Popen
 
 from lib.slurm import Slurm
@@ -139,23 +139,27 @@ class RunManager(object):
             end=end_year)
 
         atm_path = os.path.join(
-            config['global']['project_path'], 'input', 'atm')
+            config['global']['project_path'],
+            'input',
+            'atm')
         output_base_path = config['global']['output_path']
         run_scripts_path = config['global']['run_scripts_path']
+        regrid_map_name = config['global']['remap_grid_name']
+        native_map_name = config['global']['native_grid_name']
 
         regrid_output_path = os.path.join(
             output_base_path, 'pp',
-            config['global']['remap_grid_name'], 'climo',
+            regrid_map_name, 'climo',
             '{}yr'.format(year_set.length))
         if not os.path.exists(regrid_output_path):
-            os.makedirs(regrid_output_dir)
-        
-        native_output_dir = os.path.join(
+            os.makedirs(regrid_output_path)
+
+        native_output_path = os.path.join(
             output_base_path, 'pp',
-            config['global']['native_map_name'], 'climo',
+            native_map_name, 'climo',
             '{}yr'.format(year_set.length))
-        if not os.path.exists(native_output_dir):
-            os.makedirs(native_output_dir)
+        if not os.path.exists(native_output_path):
+            os.makedirs(native_output_path)
 
         if required_jobs.get('ncclimo'):
             # Add the ncclimo job to the runmanager
@@ -173,6 +177,14 @@ class RunManager(object):
                 start_year=year_set.set_start_year,
                 end_year=year_set.set_end_year,
                 _type='atm')
+
+            output_path = os.path.join(
+                output_base_path, 'pp',
+                regrid_map_name, 'monthly', 'ts',
+                '{}yr'.format(year_set.length))
+            if not os.path.exists(output_path):
+                os.makedirs(output_path)
+
             # Add the timeseries job to the runmanager
             self.add_timeseries(
                 start_year=year_set.set_start_year,
@@ -181,7 +193,7 @@ class RunManager(object):
                 input_path=atm_path,
                 regrid_map_path=config['ncclimo']['regrid_map_path'],
                 var_list=config['ncclimo']['var_list'],
-                output_path=output_base_path,
+                output_path=output_path,
                 file_list=file_list)
 
         if required_jobs.get('aprime') or required_jobs.get('aprime_diags'):
@@ -193,13 +205,34 @@ class RunManager(object):
             host_url = '/'.join([
                 config['global']['img_host_server'],
                 os.environ['USER'],
-                host_directory])
+                config['global']['experiment'],
+                'a-prime', '{start:04d}-{end:04d}'.format(
+                    start=year_set.set_start_year,
+                    end=year_set.set_end_year)])
             web_directory = os.path.join(
                 config['global']['host_directory'],
                 os.environ['USER'],
                 host_directory)
+            target_host_path = os.path.join(
+                config['global']['host_directory'],
+                os.environ['USER'],
+                config['global']['experiment'],
+                'a-prime', '{start:04d}-{end:04d}'.format(
+                    start=year_set.set_start_year,
+                    end=year_set.set_end_year))
+
+            output_path = os.path.join(
+                output_base_path, 'diags',
+                regrid_map_name, 'a-prime',
+                '{start:04d}-{end:04d}'.format(
+                    start=year_set.set_start_year,
+                    end=year_set.set_end_year))
+            if not os.path.exists(output_path):
+                os.makedirs(output_path)
 
             self.add_aprime(
+                target_host_path=target_host_path,
+                output_path=output_path,
                 web_directory=web_directory,
                 host_url=host_url,
                 start_year=year_set.set_start_year,
@@ -226,6 +259,15 @@ class RunManager(object):
                 config.get('global').get('experiment'),
                 config.get('amwg').get('host_directory'),
                 set_string])
+            
+            output_path = os.path.join(
+                output_base_path, 'diags',
+                regrid_map_name, 'amwg',
+                '{start:04d}-{end:04d}'.format(
+                    start=year_set.set_start_year,
+                    end=year_set.set_end_year))
+            if not os.path.exists(output_path):
+                os.makedirs(output_path)
 
             self.add_amwg(
                 start_year=year_set.set_start_year,
@@ -234,8 +276,8 @@ class RunManager(object):
                 resource_path=config['global']['resource_dir'],
                 web_directory=web_directory,
                 host_url=host_url,
-                output_path=output_base_path,
-                regrid_path=regrid_output_dir,
+                output_path=output_path,
+                regrid_path=regrid_output_path,
                 diag_home=config['amwg']['diag_home'])
 
         if required_jobs.get('e3sm_diags') or required_jobs.get('acme_diags'):
@@ -252,8 +294,18 @@ class RunManager(object):
                 config['global']['experiment'],
                 config['e3sm_diags']['host_directory'],
                 set_string])
+            
+            output_path = os.path.join(
+                output_base_path, 'diags',
+                regrid_map_name, 'e3sm_diags',
+                '{start:04d}-{end:04d}'.format(
+                    start=year_set.set_start_year,
+                    end=year_set.set_end_year))
+            if not os.path.exists(output_path):
+                os.makedirs(output_path)
 
             self.add_e3sm(
+                regrid_output_path=regrid_output_path,
                 start_year=year_set.set_start_year,
                 end_year=year_set.set_end_year,
                 year_set=year_set,
@@ -261,7 +313,7 @@ class RunManager(object):
                 web_directory=web_directory,
                 host_url=host_url,
                 reference_data_path=config['e3sm_diags']['reference_data_path'],
-                output_path=output_base_path,
+                output_path=output_path,
                 seasons=config['e3sm_diags']['seasons'],
                 backend=config['e3sm_diags']['backend'],
                 sets=config['e3sm_diags']['sets'])
@@ -321,25 +373,19 @@ class RunManager(object):
             input_path (str): the path to the raw cam.h0 files
             regrid_map_path (str): the path to the regrid map
             var_list (list(str)): the list of variables to extract
+            output_path (str): the path to store the timeseries output
         """
         start_year = kwargs['start_year']
         end_year = kwargs['end_year']
         year_set = kwargs['year_set']
         input_path = kwargs['input_path']
+        output_path = kwargs['output_path']
         regrid_map_path = kwargs['regrid_map_path']
         var_list = kwargs['var_list']
         file_list = kwargs['file_list']
 
         if not self._precheck(year_set, 'timeseries'):
             return
-
-        set_length = end_year - start_year + 1
-        timeseries_output_dir = os.path.join(
-            self.output_path,
-            'monthly',
-            '{}yr'.format(set_length))
-        if not os.path.exists(timeseries_output_dir):
-            os.makedirs(timeseries_output_dir)
 
         config = {
             'ui': self.ui,
@@ -351,7 +397,7 @@ class RunManager(object):
             'var_list': var_list,
             'start_year': start_year,
             'end_year': end_year,
-            'output_directory': timeseries_output_dir,
+            'output_directory': output_path,
             'regrid_map_path': regrid_map_path
         }
         timeseries = Timeseries(
@@ -367,6 +413,7 @@ class RunManager(object):
         Add an APrime job to the job_list
 
         Parameters:
+            output_path (str): The path to store the output
             web_directory (str): The path to the directory to store images for hosting
             host_url (str): The url to access the images once their hosted
             year_set (YearSet): The YearSet that this job belongs to
@@ -377,13 +424,16 @@ class RunManager(object):
             test_atm_res (str): the atm resolution for the test input
             test_mpas_mesh_name (str): The test mpas mesh name
             aprime_code_path (str): the path to the aprime code
+            target_host_path (str): the real hosting directory
         """
+        target_host_path = kwargs['target_host_path']
         web_directory = kwargs['web_directory']
         host_url = kwargs['host_url']
         year_set = kwargs['year_set']
         start_year = kwargs['start_year']
         end_year = kwargs['end_year']
         input_base_path = kwargs['input_base_path']
+        output_path = kwargs['output_path']
         resource_path = kwargs['resource_path']
         test_atm_res = kwargs['test_atm_res']
         test_mpas_mesh_name = kwargs['test_mpas_mesh_name']
@@ -397,10 +447,7 @@ class RunManager(object):
             start=start_year,
             end=end_year)
 
-        project_dir = os.path.join(
-            self.output_path,
-            'aprime_diags',
-            year_set_string)
+        project_dir = output_path
         if not os.path.exists(project_dir):
             os.makedirs(project_dir)
 
@@ -427,6 +474,7 @@ class RunManager(object):
             return
 
         config = {
+            'target_host_path': target_host_path,
             'ui': self.ui,
             'web_dir': web_directory,
             'host_url': host_url,
@@ -476,6 +524,7 @@ class RunManager(object):
         host_url = kwargs['host_url']
         regrid_path = kwargs['regrid_path']
         diag_home = kwargs['diag_home']
+        output_path = kwargs['output_path']
 
         if not self._precheck(year_set, 'amwg'):
             return
@@ -483,14 +532,6 @@ class RunManager(object):
         year_set_string = '{start:04d}-{end:04d}'.format(
             start=start_year,
             end=end_year)
-
-        # Setup the amwg output directory
-        output_path = os.path.join(
-            self.output_path,
-            'amwg_diag',
-            year_set_string)
-        if not os.path.exists(output_path):
-            os.makedirs(output_path)
 
         # Setup the AMWG temp directory
         temp_path = os.path.join(
@@ -562,6 +603,8 @@ class RunManager(object):
         backend = kwargs['backend']
         sets = kwargs['sets']
         resource_path = kwargs['resource_path']
+        output_path = kwargs['output_path']
+        regrid_output_path = kwargs['regrid_output_path']
 
         if not self._precheck(year_set, 'e3sm_diags'):
             msg = 'rejecting e3sm_diags'
@@ -575,13 +618,6 @@ class RunManager(object):
             start=start_year,
             end=end_year)
 
-        # Setup output directory
-        output_path = os.path.join(
-            self.output_path,
-            'e3sm_diags',
-            set_string)
-        if not os.path.exists(output_path):
-            os.makedirs(output_path)
         # Setup temp directory
         temp_path = os.path.join(
             self.output_path,
@@ -601,13 +637,9 @@ class RunManager(object):
             logging.error(msg)
             return
 
-        regrid_path = os.path.join(
-            self.output_path,
-            'climo_regrid')
-
         config = {
             'ui': self.ui,
-            'regrid_base_path': regrid_path,
+            'regrid_output_path': regrid_output_path,
             'web_dir': web_directory,
             'host_url': host_url,
             'experiment': self.caseID,
@@ -628,7 +660,7 @@ class RunManager(object):
         e3sm_diag = E3SMDiags(
             config=config,
             event_list=self.event_list)
-        msg = "Adding ACME Diagnostic to the job list: {}".format(
+        msg = "Adding E3SM Diagnostic to the job list: {}".format(
             str(e3sm_diag))
         logging.info(msg)
         year_set.add_job(e3sm_diag)
@@ -657,7 +689,7 @@ class RunManager(object):
 
         for job_set in self.job_sets:
             # if the job state is ready, but hasnt started yet
-            if job_set.status in [SetStatus.DATA_READY, SetStatus.RUNNING]:
+            if job_set.status in [SetStatus.DATA_READY, SetStatus.RUNNING, SetStatus.FAILED]:
                 # Iterate over all jobs in the set
                 for job in job_set.jobs:
                     if job.status == JobStatus.INVALID:
@@ -688,43 +720,47 @@ class RunManager(object):
                         continue
                     # If the job is valid, start it
                     if job.status == JobStatus.VALID:
-                        if not self.check_max_running_jobs():
-                            slurm = Slurm()
-                            msg = "Submitting {job}".format(job=job.type)
+                        if self.check_max_running_jobs():
+                            return
+                        slurm = Slurm()
+                        msg = "Submitting {job}-{start:04d}-{end:04d}".format(
+                            job=job.type,
+                            start=job.start_year,
+                            end=job.end_year)
+                        print_line(
+                            ui=self.ui,
+                            line=msg,
+                            event_list=self.event_list,
+                            current_state=True,
+                            ignore_text=True)
+                        status = job.execute(dryrun=self._dryrun)
+                        if status == -1:
+                            continue
+                        if job.job_id == 0:
+                            msg = '{job}-{start:04d}-{end:04d} precomputed, skipping'.format(
+                                job=job.type,
+                                start=job.start_year,
+                                end=job.end_year)
                             print_line(
                                 ui=self.ui,
                                 line=msg,
-                                event_list=self.event_list,
-                                current_state=True,
-                                ignore_text=True)
-                            status = job.execute(dryrun=self._dryrun)
-                            if status == -1:
-                                continue
-                            if job.job_id == 0:
-                                msg = '{job}-{start:04d}-{end:04d} precomputed, skipping'.format(
-                                    job=job.type,
-                                    start=job.start_year,
-                                    end=job.end_year)
-                                print_line(
-                                    ui=self.ui,
-                                    line=msg,
-                                    event_list=self.event_list)
-                                logging.info(msg)
-                                continue
-                            try:
-                                slurm.showjob(job.job_id)
-                            except:
-                                msg = "Error submitting {job} to queue".format(job=job.type)
-                                print_line(
-                                    ui=self.ui,
-                                    line=msg,
-                                    event_list=self.event_list)
-                                job.status = JobStatus.VALID
-                                job.status = JobStatus.VALID
-                                return
-                            else:
-                                self.running_jobs.append(job)
-                                self.monitor_running_jobs()
+                                event_list=self.event_list)
+                            logging.info(msg)
+                            self.handle_completed_job(job)
+                            continue
+                        try:
+                            slurm.showjob(job.job_id)
+                        except:
+                            msg = "Error submitting {job} to queue".format(job=job.type)
+                            print_line(
+                                ui=self.ui,
+                                line=msg,
+                                event_list=self.event_list)
+                            job.status = JobStatus.VALID
+                            return
+                        else:
+                            self.running_jobs.append(job)
+                            self.monitor_running_jobs()
 
     def monitor_running_jobs(self):
         msg = 'Updating job list'
@@ -810,7 +846,7 @@ class RunManager(object):
             ui=self.ui,
             line=msg,
             event_list=self.event_list,
-            current_state=True)
+            current_state=False)
         job_set = None
         for s in self.job_sets:
             if s.set_number == job.year_set:
@@ -856,6 +892,18 @@ class RunManager(object):
                     break
             head, _ = os.path.split(host_dir)
             os.chmod(head, 0755)
+
+            # move the files from the place that aprime auto
+            # generates them to where we actually want them to be
+            target_host_dir = job.config['target_host_path']
+            head, tail = os.path.split(target_host_dir)
+            if not os.path.exists(head):
+                os.makedirs(head)
+            if os.path.exists(host_dir) and not os.path.exists(target_host_dir):
+                move(
+                    src=host_dir,
+                    dst=target_host_dir)
+
             msg = '{job} hosted at {url}/index.html'.format(
                 url=job.config.get('host_url'),
                 job=job.type)
@@ -884,8 +932,11 @@ class RunManager(object):
         elif job.type == 'e3sm_diags':
             img_src = job.config.get('results_dir')
             self.setup_local_hosting(job, img_src)
-            msg = '{job} hosted at {url}/viewer/index.html'.format(
-                url=job.config.get('host_url'),
+            host_url = '{url}/viewer/index.html'.format(
+                url=job.config.get('host_url'))
+            job.config['host_url'] = host_url
+            msg = '{job} hosted at {url}'.format(
+                url=host_url,
                 job=job.type)
             print_line(
                 ui=self.ui,
@@ -934,7 +985,8 @@ class RunManager(object):
                 current_state=True,
                 ignore_text=False)
             logging.info(msg)
-            copytree(src=img_src, dst=host_dir)
+            if os.path.exists(img_src) and not os.path.exists(host_dir):
+                copytree(src=img_src, dst=host_dir)
 
             msg = 'Fixing permissions for {dir}'.format(
                 dir=host_dir)
